Create Task-4.md
Opened the AWS console and primarily changed the region to the south and started launching the instances in EC2, downloaded the key pair and kept it safe. After running the instance, i have given connect to the RDP client and retrieve the username and password for the RDP connection. Opened the RDP from my windows and given the public IP and connected. As per the task given, i have opened the cmd in the remote windows instance.